Implement bottom siphoning
This *should* cover all cover cases but the code (and algorithm) is a
bit messy so this needs to be tested more.

Left siphoning could be implemented easily by copy-pasting the code and
changing all coordinates (x, y) <-> (y, x), but I want to try and do it
in a bit more of a generic manner.