Implement siphoning in both directions
I mostly believe that this works for all input shapes without a parity
cube. I know of at least one bug when a parity cube is present: the
input shape

        x
    x x x
    x x x

still crashes because it doesn't move the parity cube away.

I guess this should be fixed by doing a siphon in the other direction
instead.