Fix bridge capacity computation
When computing the capacity of a connector, it would also include the
rest of the chunk even though that is not how it is described in the
paper.

Also, removed the default configuration (because it also gets added when
loading a JSON string, and is generally useless), and fixed the
components computation when the configuration is empty.