Do gathering in a smarter way
Determine beforehand where the cube will move to, and pick
counter-clockwise or clockwise movement depending on which one will not
pass along the origin.

This allows us to avoid leaving the configuration bounding box. We also
updated compaction not to leave the bounding box, and disabled
canonicalization. This way, we now ensure (hopefully) that the bounding
box is respected.