Check square stability when finding chain moves
This is necessary because there is a special case where the chain move
considered could try to move the connector of the chunk. This can happen
for vertical chain moves when the connector is at the top-left corner of
the chunk, or for horizontal chain moves when the connector is at the
bottom-right corner. This used to cause a crash as preserveChunkiness()
assumes this should not happen and throws.

In these cases a chain move is simply not possible, so we return null
and this adequately solves the issue.